requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.1.1
python-dotenv==1.2.2
python-telegram-bot==22.7
reportlab==4.4.10
